Split up StoryGenerator and add StoryFragment loading from file
- NOTE: Generated Scenario looks a bit odd and does not play out properly. Need to bisect!
